# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neryva Agent Studio",
    "description": "Governance and control layer for LLM applications",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/harness/sessions": {
      "get": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "List harness sessions (operator only)",
        "operationId": "list_harness_sessions_api_v1_harness_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Harness Sessions Api V1 Harness Sessions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "Create a harness session (operator only)",
        "operationId": "create_harness_session_api_v1_harness_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Harness Session Api V1 Harness Sessions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/harness/sessions/{session_id}": {
      "get": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "Get a harness session",
        "operationId": "get_harness_session_api_v1_harness_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Harness Session Api V1 Harness Sessions  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "Delete a harness session",
        "operationId": "delete_harness_session_api_v1_harness_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/harness/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "Send a message in a harness session",
        "operationId": "send_harness_message_api_v1_harness_sessions__session_id__messages_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessSendMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Send Harness Message Api V1 Harness Sessions  Session Id  Messages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/harness/sessions/{session_id}/fork": {
      "post": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "Fork a harness session at a message boundary",
        "description": "Fork any thread at a message boundary (Arch \u00a77.1, P7-5).\n\nCreates a new active session that copies the original's history up to the\nfork point. The original session's ID and history stay unchanged; both\ncan then be resumed independently. This implements the same pattern as\nOpenCode's fork and Claude Agent SDK's ``fork_session``.",
        "operationId": "fork_harness_session_api_v1_harness_sessions__session_id__fork_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessForkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Fork Harness Session Api V1 Harness Sessions  Session Id  Fork Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/harness/sessions/{session_id}/replay": {
      "post": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "Replay a session with a different model/provider",
        "description": "Replay all user turns from an existing session through a new model.\n\nCreates a new session with the same user messages, re-running each\nthrough the specified provider/model. Used for model comparison via\ncontrolled replay (Arch \u00a713, P7-5 redacted-trace replay).",
        "operationId": "replay_harness_session_api_v1_harness_sessions__session_id__replay_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessReplayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Replay Harness Session Api V1 Harness Sessions  Session Id  Replay Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/harness/compare": {
      "post": {
        "tags": [
          "Harness (Internal Only)"
        ],
        "summary": "Run a single prompt against multiple models simultaneously",
        "description": "Execute a comparison prompt against multiple provider/model pairs.\n\nResults are collected in parallel and returned together for side-by-side\ncomparison. All calls are tagged internal (AGENTS.md boundary).",
        "operationId": "run_model_comparison_api_v1_harness_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarnessCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Model Comparison Api V1 Harness Compare Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HarnessCompareRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "models": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "minItems": 1,
            "title": "Models",
            "description": "[{provider, model}, ...]"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "models"
        ],
        "title": "HarnessCompareRequest"
      },
      "HarnessForkRequest": {
        "properties": {
          "at_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At Message Id"
          }
        },
        "type": "object",
        "title": "HarnessForkRequest"
      },
      "HarnessReplayRequest": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "title": "HarnessReplayRequest"
      },
      "HarnessSendMessage": {
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "HarnessSendMessage"
      },
      "HarnessSessionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "One of: provider_test, model_comparison, prompt_ab, fork_replay, adversarial_sweep"
          },
          "provider": {
            "type": "string",
            "minLength": 1,
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "minLength": 1,
            "title": "Model"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "provider",
          "model"
        ],
        "title": "HarnessSessionCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}